### 2024-11-19 #7504 @editorialbot
[PRE REVIEW]: ggret: An R package for visualising and manipulating tree‑based phylogenetic networks
**Submitting author:** @grdspcht (<a href="http://orcid.org/0009-0001-3320-0557">Gerd Specht</a>)
**Repository:** https://github.com/grdspcht/ggret/
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
