### 2024-11-25 #7504 @Kevin-Mattheus-Moerman
- [ ] Your project appears to lack an OSI approved open source software license. Please add one in the form of a plain text LICENSE file.
