### 2025-01-12 b06f5c2
Addressing JOSS comments
