+++ paper/paper.md
@@ -19,7 +19,7 @@ authors:
     affiliation: "1,2"
     corresponding: true
   - name: Clemens Schmid
-    orcid: 0000-0003-3448-571
+    orcid: 0000-0003-3448-5715
     affiliation: "3"
   - name: Denise Kühnert
     orcid: 0000-0002-5657-018X
