+++ paper/paper.bib
@@ -1,28 +1,3 @@
-
-@book{noauthor_notitle_nodate,
-}
-
-@misc{noauthor_notitle_nodate-1,
-	url = {https://www.stat.auckland.ac.nz/~paul/RGraphics/rgraphics.html},
-	urldate = {2023-07-03},
-}
-
-@misc{noauthor_r_nodate,
-	title = {R {Graphics}, {Third} {Edition}},
-	url = {https://www.routledge.com/R-Graphics-Third-Edition/Murrell/p/book/9780367780692},
-	abstract = {This third edition of Paul Murrell’s classic book on using R for graphics represents a major update, with a complete overhaul in focus and scope. It focuses primarily on the two core graphics packages in R - graphics and grid - and has a new section on integrating graphics. This section includes three new chapters: importing external images in to R; integrating the graphics and grid systems; and advanced SVG graphics.The emphasis in this third edition is on having the ability to produce detail},
-	language = {en},
-	urldate = {2023-07-03},
-	journal = {Routledge \& CRC Press},
-	file = {Snapshot:/home/user/Zotero/storage/A9PZMWZN/9780367780692.html:text/html},
-}
-
-@misc{noauthor_notitle_nodate-2,
-	url = {https://www.stat.auckland.ac.nz/~paul/RG2e/},
-	urldate = {2023-07-03},
-	file = {stat.auckland.ac.nz/~paul/RG2e/:/home/user/Zotero/storage/LPUEIET6/RG2e.html:text/html},
-}
-
 @book{murrell_r_2005,
 	address = {New York},
 	edition = {3},
@@ -36,18 +11,6 @@
 	doi = {10.1201/9781420035025},
 }
 
-@book{murrell_r_2018,
-	title = {R {Graphics}},
-	isbn = {978-1-4398-3177-9},
-	abstract = {Extensively updated to reflect the evolution of statistics and computing, the second edition of the bestselling R Graphics comes complete with new packages and new examples. Paul Murrell, widely known as the leading expert on R graphics, has developed an in-depth resource that helps both neophyte and seasoned users master the intricacies of R graph},
-	language = {en},
-	publisher = {CRC Press},
-	author = {Murrell, Paul},
-	month = nov,
-	year = {2018},
-	note = {Google-Books-ID: ol3RBQAAQBAJ},
-	keywords = {Computers / Software Development \& Engineering / Computer Graphics, Mathematics / Probability \& Statistics / General},
-}
 
 @book{wickham_ggplot2_2016,
 	address = {Cham},
@@ -221,6 +184,7 @@
 	volume = {4},
 	issn = {1664-8021},
 	url = {https://www.frontiersin.org/articles/10.3389/fgene.2013.00206},
+	doi = {10.3389/fgene.2013.00206},
 	urldate = {2023-07-04},
 	journal = {Frontiers in Genetics},
 	author = {Arenas, Miguel},
@@ -231,6 +195,7 @@
 @article{kuhner_graphml_2013,
 	title = {{GraphML} specializations to codify ancestral recombinant graphs},
 	volume = {4},
+	doi = {10.3389/fgene.2013.00146},
 	issn = {1664-8021},
 	url = {https://www.frontiersin.org/articles/10.3389/fgene.2013.00146},
[…]
