### 2025-01-12 #7504 @grdspcht
@Kevin-Mattheus-Moerman please excuse my late response. A recent job change and end of year activities made me a bit forgetful. I have addressed the bibfile issues and the contribution guidelines. Tomorrow I will address the licensing issue so the MIT license is correctly applied to this project.
