### 2025-01-14 #7504 @grdspcht
@Kevin-Mattheus-Moerman Added the MIT license files.
