### 2025-03-03 #7773 @Gammerdinger
I have completed my review.

**Do the authors describe how this software compares to other commonly-used packages?**
I could not find a discussion of the State of the Field in the manuscript.

**Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?**
The ?ggret is nicely done and useful. It would be nice if there was more documentation on the GitHub.

**Are there automated tests or manual steps described so that the functionality of the software can be verified?**
There is a vignette that works if that is sufficient.
