### 2025-03-09 #7773 @shrestha-c
Hello, I have completed my review.
On the **Software Paper** section - "**State of the field: Do the authors describe how this software compares to other commonly-used packages?**"
While the paper effectively outlines the functionalities and potential benefits of "ggret," the review lacks a comprehensive "State of the Field" section that compares "ggret" to other existing phylogenetic analysis tools.
