### 2025-03-10 #7773 @lpantano
Hi @grdspcht,

To increase automation you could implement this: https://usethis.r-lib.org/reference/use_pkgdown.html

Let me know if you have any question,

Thanks
