### 2025-03-14 d5a92b7
Update paper.md
Completed Statement of need section with a mention of existing tools
