### 2025-03-16 #7773 @grdspcht
@lpantano
Dear editors,
thank you for your reviews and your suggestions. I am currently testing pkgdown and will most likely start hosting documentation with it. I will also provide a state of the field section. I am aware one other package that is similar to ggret.
I will provide the aformentioned changes by the end of the week.
