### 2025-03-26 c2446fc
Update paper.md
typo fixes
