+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://grdspcht.github.io/ggret/
+template:
+  bootstrap: 5
+
