+++ pkgdown.yml
@@ -1,0 +1,9 @@
+pandoc: 3.1.11
+pkgdown: 2.1.1
+pkgdown_sha: ~
+articles:
+  intro_to_ggret: intro_to_ggret.html
+last_built: 2025-03-31T21:57Z
+urls:
+  reference: https://grdspcht.github.io/ggret/reference
+  article: https://grdspcht.github.io/ggret/articles
