### 2025-04-03 #7773 @grdspcht
@lpantano
We have updated the manuscript with a paragraph on the state of the field and we have also set up a pkgdown site at:  https://grdspcht.github.io/ggret/
