### 2025-04-14 #7773 @lpantano
Hello @grdspcht, I think this is a pretty good shape. The last modification will be to shorten the paper to 3 pages. Do you think you can make a panel of figures and the code can be left to the documentation? or add one example only with code? Let me know. Thanks for your patience.
