### 2025-04-30 #7773 @lpantano
Hello @grdspcht , any chance that the paper could be reduced in size, everything is ready to go after that. Thanks!
