### 2025-05-01 #7773 @grdspcht
@lpantano I am planning to reduce the length of the paper but this might take a while since work obligations are keeping me very busy right now. Please excuse the delay. I'll try to get to this as soon as possible.
