### 2025-05-20 #7773 @arthurKocher
Dear @lpantano, we have now shortened the paper by combining the plots into one figure and referring to the documentation of the package for usage details. This should fit in 3 pages. Please let us know if anything else is required.
