+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
   intro_to_ggret: intro_to_ggret.html
-last_built: 2025-04-01T16:10Z
+last_built: 2025-05-20T09:03Z
 urls:
   reference: https://grdspcht.github.io/ggret/reference
   article: https://grdspcht.github.io/ggret/articles
