### 2025-05-23 068b543
Update paper.md
Corrected typo in figure legend
