+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
   intro_to_ggret: intro_to_ggret.html
-last_built: 2025-05-20T09:03Z
+last_built: 2025-05-23T08:15Z
 urls:
   reference: https://grdspcht.github.io/ggret/reference
   article: https://grdspcht.github.io/ggret/articles
