### 2025-05-23 #7773 @grdspcht
@lpantano
We have:
1. double checked authors and their associated ORCIDs
2. published a release on GitHub ([v.1.0.0](https://github.com/grdspcht/ggret/releases/tag/v.1.0.0))
3. archived the release on Zenodo ([DOI: 10.5281/zenodo.15495857](https://doi.org/10.5281/zenodo.15495857))
4. made sure that title and authors match those in the JOSS paper
5. matched archive and software licenses (MIT)
