### 2025-06-07 #7773 @Kevin-Mattheus-Moerman
- [ ] Words like `visualisation` in the text and title are British English. However `visualizing` / `visualization` are American English. Please edit to use one form consistently.
