### 2025-06-10 9e67408
Fixing Figure 1
